+++ app/vmui/packages/vmui/src/components/Configurators/QueryEditor/QueryEditor.tsx
@@ -1,8 +1,7 @@
-import React, { FC, useEffect, useRef, useState } from "preact/compat";
+import React, { FC, Ref, useEffect, useRef, useState } from "preact/compat";
 import { KeyboardEvent } from "react";
 import { ErrorTypes } from "../../../types";
 import TextField from "../../Main/TextField/TextField";
-import QueryEditorAutocomplete from "./QueryEditorAutocomplete";
 import "./style.scss";
 import { QueryStats } from "../../../api/types";
 import { partialWarning, seriesFetchedWarning } from "./warningText";
@@ -10,6 +9,15 @@ import { AutocompleteOptions } from "../../Main/Autocomplete/Autocomplete";
 import useDeviceDetect from "../../../hooks/useDeviceDetect";
 import { useQueryState } from "../../../state/query/QueryStateContext";
 import debounce from "lodash.debounce";
+
+export interface QueryEditorAutocompleteProps {
+  value: string;
+  anchorEl: Ref<HTMLInputElement>;
+  caretPosition: [number, number]; // [start, end]
+  hasHelperText: boolean;
+  onSelect: (val: string, caretPosition: number) => void;
+  onFoundOptions: (val: AutocompleteOptions[]) => void;
+}
 
 export interface QueryEditorProps {
   onChange: (query: string) => void;
@@ -19,6 +27,7 @@ export interface QueryEditorProps {
   value: string;
   oneLiner?: boolean;
   autocomplete: boolean;
+  autocompleteEl?: FC<QueryEditorAutocompleteProps>;
   error?: ErrorTypes | string;
   stats?: QueryStats;
   label: string;
@@ -32,6 +41,7 @@ const QueryEditor: FC<QueryEditorProps> = ({
   onArrowUp,
   onArrowDown,
   autocomplete,
+  autocompleteEl: AutocompleteEl,
   error,
   stats,
   label,
@@ -44,7 +54,7 @@ const QueryEditor: FC<QueryEditorProps> = ({
   const [caretPosition, setCaretPosition] = useState<[number, number]>([0, 0]);
   const autocompleteAnchorEl = useRef<HTMLInputElement>(null);
 
-  const [showAutocomplete, setShowAutocomplete] = useState(autocomplete);
+  const [showAutocomplete, setShowAutocomplete] = useState(!!AutocompleteEl);
   const debouncedSetShowAutocomplete = useRef(debounce(setShowAutocomplete, 500)).current;
 
   const warning = [
@@ -110,7 +120,7 @@ const QueryEditor: FC<QueryEditorProps> = ({
   };
 
   useEffect(() => {
-    setOpenAutocomplete(autocomplete);
+    setOpenAutocomplete(!!AutocompleteEl);
   }, [autocompleteQuick]);
 
   useEffect(() => {
@@ -137,8 +147,8 @@ const QueryEditor: FC<QueryEditorProps> = ({
         inputmode={"search"}
         caretPosition={caretPosition}
       />
-      {showAutocomplete && autocomplete && (
-        <QueryEditorAutocomplete
+      {showAutocomplete && autocomplete && AutocompleteEl && (
+        <AutocompleteEl
           value={value}
           anchorEl={autocompleteAnchorEl}
           caretPosition={caretPosition}